Refactor Detection-to-VLM GenAI Example (#287)
* fix(genai): simplify detection VLM config

* fix(genai): improve detector app output

* docs(genai): clarify VLM model setup
