+++ examples/genai/detection-to-vlm-assistant/README.md
@@ -5,14 +5,14 @@
 | --- | --- |
 | Category | genai |
 | Difficulty | Advanced |
-| Tags | object-detection, genai, yolo26, rtsp, insight, openai-compatible |
+| Tags | object-detection, genai, yolo26, rtsp, insight, vlm |
 | Languages | Python |
 | Status | experimental |
 | Binary Name | detection-to-vlm-assistant |
 | Model | yolo26m-det-bf16-mla_tess-b1 |
 
 ## Concept
-This example decodes an RTSP stream, runs YOLO26 detection with internal box decode, and sends video plus object-detection metadata to Insight. Insight video uses `VideoSender`, which owns raw-frame caps, conversion, H.264 encoding, RTP packetization, and UDP output. When OpenAI is enabled, the highest-score detected person is cropped and sent to the configured OpenAI-compatible Gemma server from a bounded background worker, so the detection and Insight loop keeps running.
+This example decodes an RTSP stream, runs YOLO26 detection with internal box decode, and sends video plus object-detection metadata to Insight. Insight video uses `VideoSender`, which owns raw-frame caps, conversion, H.264 encoding, RTP packetization, and UDP output. When GenAI is enabled, the highest-score detected person is cropped and sent to the configured GenAI server from a bounded background worker, so the detection and Insight loop keeps running.
 
 ## Preview
 Detection metadata visualized in Insight:
@@ -42,9 +42,11 @@ Use the same `neat` output to set `insight.host`, `video_port`, and `metadata_po
 ## Prerequisites
 - Installed Neat Development Environment + Neat Library.
 - Model artifacts are user-managed and should be downloaded into `assets/models/`. Download the default YOLO26 detector model, or set `model.path` to another readable model package.
+- Choose a vision-language model from the [SiMa.ai VLM collection](https://huggingface.co/collections/simaai/vision-language-models), download it locally, and set `genai_server.model.path` to that model directory.
+- Set `genai_server.model.name` to the served model name used by the detector app.
 - RTSP source created in Insight or provided by your camera.
 - Insight receiver running at the configured host and ports.
-- OpenAI-compatible Gemma server running when `openai.enabled` is true.
+- GenAI server running when `genai.enabled` is true.
 
 ## Get The Apps Repo
 Use the [Neat Development Environment](https://developer.sima.ai/software/getting-started/dev-environment/) with the [Neat Library](https://developer.sima.ai/software/getting-started/neat-library/) installed for setup and compilation.
@@ -92,24 +94,43 @@ insight:
   video_port: <videoUDP start port from neat>              # UDP video port.
   metadata_port: <metadataUDP start port from neat>        # UDP metadata port.
 
-openai:
+genai_server:
+  host: 0.0.0.0                                            # GenAI server bind host.
+  port: 9998                                               # GenAI server port.
+  model:
+    name: <served-vlm-model-name>                          # Model name used by requests.
+    path: <path-to-vlm-model-dir>                          # Local VLM model directory.
+
+genai:
   enabled: false                                           # Disable for detection plus Insight only.
+  host: 127.0.0.1                                          # GenAI server client host.
+  port: 9998                                               # GenAI server client port.
+  system_prompt: <system-prompt>
+  user_prompt: <user-prompt>
 ```
+
+Change `genai.system_prompt` and `genai.user_prompt` to influence the GenAI output. The default asks what the detected person is doing in the crop.
 
 ## Run
 From the `apps` repository root:
 
+In one terminal, start the local GenAI server:
+
 ```bash
-python3 examples/genai/detection-to-vlm-assistant/src/python/main.py \
+python3 examples/genai/detection-to-vlm-assistant/src/python/genai_server.py \
   --config examples/genai/detection-to-vlm-assistant/src/common/config.yaml
 ```
 
-Set `openai.enabled: false` in the config to run only the detection and Insight path.
+In a second terminal, start the detection-to-Insight pipeline:
 
-For local model comparison, example configs are available under
-`sandbox/configs/detection-to-vlm-assistant/<model-name>/config.yaml`.
+```bash
+python3 examples/genai/detection-to-vlm-assistant/src/python/detector_app.py \
+  --config examples/genai/detection-to-vlm-assistant/src/common/config.yaml
+```
 
-The OpenAI path checks `/v1/models` before sending a crop, waits at least `openai.interval_seconds` between attempts, and keeps at most `openai.max_pending_requests` queued or in-flight requests.
+Set `genai.enabled: false` in the config to run only the detection and Insight path.
+
+The GenAI path checks `/v1/models` before sending a crop, waits at least `genai.interval_seconds` between attempts, and keeps at most `genai.max_pending_requests` queued or in-flight requests.
 
 ## Appendix: Additional Models
 Other supported batch-1 YOLO26 detector models:
@@ -123,5 +144,5 @@ Other supported batch-1 YOLO26 detector models:
 Replace the default filename in the download command and `model.path`.
 
 ## Source Files
-- Python source: `src/python/main.py`, `src/python/utils/helpers.py`, `src/python/utils/openai_commenter.py`
+- Python source: `src/python/detector_app.py`, `src/python/genai_server.py`
 - Shared assets: `src/common/`
