--- examples/genai/detection-to-vlm-assistant/src/python/utils/openai_commenter.py
@@ -1,105 +1,0 @@
-"""Bounded background OpenAI requests for detected people."""
-
-from __future__ import annotations
-
-from queue import Empty, Full, Queue
-import threading
-import time
-from urllib import error, request
-
-import numpy as np
-
-from .helpers import (
-    Config,
-    best_box_for_label,
-    crop_box,
-    request_person_crop_comment,
-)
-
-
-class OpenAICommenter:
-    def __init__(self, cfg: Config, labels: list[str]):
-        self.cfg = cfg
-        self.labels = labels
-        self.queue: Queue[np.ndarray] = Queue(maxsize=cfg.openai_max_pending_requests)
-        self.stop_event = threading.Event()
-        self.worker = threading.Thread(target=self._run, daemon=True)
-        self.lock = threading.Lock()
-        self.last_enqueue_at = 0.0
-        self.in_flight = False
-        self.started = False
-
-    def start(self) -> None:
-        if self.cfg.openai_enabled and not self.started:
-            self.worker.start()
-            self.started = True
-
-    def try_enqueue(self, frame: np.ndarray, boxes: list[dict]) -> None:
-        if not self.cfg.openai_enabled:
-            return
-        now = time.monotonic()
-        if now - self.last_enqueue_at < self.cfg.openai_interval_seconds:
-            return
-        if self._pending_count() >= self.cfg.openai_max_pending_requests:
-            print("openai: queue busy, dropping request", flush=True)
-            self.last_enqueue_at = now
-            return
-
-        box = best_box_for_label(boxes, self.labels, "person")
-        if box is None:
-            return
-
-        try:
-            self.queue.put_nowait(crop_box(frame, box).copy())
-            self.last_enqueue_at = now
-        except Full:
-            print("openai: queue full, dropping request", flush=True)
-            self.last_enqueue_at = now
-
-    def close(self) -> None:
-        self.stop_event.set()
-        if self.started:
-            self.worker.join(timeout=1.0)
-
-    def _pending_count(self) -> int:
-        with self.lock:
-            return self.queue.qsize() + int(self.in_flight)
-
-    def _set_in_flight(self, value: bool) -> None:
-        with self.lock:
-            self.in_flight = value
-
-    def _run(self) -> None:
-        while not self.stop_event.is_set():
-            try:
-                crop = self.queue.get(timeout=0.2)
-            except Empty:
-                continue
-
-            self._set_in_flight(True)
-            try:
-                if self._server_ready():
-                    response = request_person_crop_comment(crop, self.cfg)
-                    if response:
-                        print(f"openai: {response}", flush=True)
-            except (TimeoutError, OSError, error.URLError) as exc:
-                print(f"openai unavailable: {exc}", flush=True)
-            except Exception as exc:
-                print(f"openai request failed: {exc}", flush=True)
-            finally:
-                self._set_in_flight(False)
-                self.queue.task_done()
-
-    def _server_ready(self) -> bool:
-        url = f"http://{self.cfg.openai_host}:{self.cfg.openai_port}/v1/models"
-        try:
-            timeout = min(self.cfg.openai_timeout_seconds, 5.0)
-            with request.urlopen(url, timeout=timeout) as res:
-                return 200 <= res.status < 300
-        except (TimeoutError, OSError, error.URLError) as exc:
-            print(
-                f"openai unavailable: could not access server at "
-                f"{self.cfg.openai_host}:{self.cfg.openai_port}: {exc}",
-                flush=True,
-            )
-            return False
