+++ examples/genai/detection-to-vlm-assistant/tests/test-scope.yaml
@@ -3,7 +3,7 @@ models:
     source: url
     url: https://docs.sima.ai/pkg_downloads/SDK{platform_version}/models/modalix/yolo26-detection/yolo26m-det-bf16-mla_tess-b1.tar.gz
     file: yolo26m-det-bf16-mla_tess-b1.tar.gz
-  gemma4-openai-server:
+  vlm-genai-server:
     source: huggingface
     repo: TBD
     path: TBD
